Update room models

+++ api/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404
-from rest_framework import status, views
+from rest_framework import exceptions, status, views
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
 from .models import MyUser, Room
 from .serializers import UserSerializer, RoomSerializer
@@ -42,6 +43,7 @@ class UserRetrieveUpdateDestroyAPIView(views.APIView):
     """
     ユーザモデルの取得(詳細)・更新・一部更新・削除APIクラス
     """
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get(self, request, pk, *args, **kwargs):
         """ユーザモデルの取得(詳細)APIに対応するハンドラメソッド"""
@@ -59,6 +61,8 @@ class UserRetrieveUpdateDestroyAPIView(views.APIView):
 
         # モデルオブジェクトを取得
         user = get_object_or_404(MyUser, pk=pk)
+        if user != request.user:
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # シリアライザオブジェクトを作成
         serializer = UserSerializer(instance=user, data=request.data)
         # バリデーション
@@ -75,6 +79,8 @@ class UserRetrieveUpdateDestroyAPIView(views.APIView):
 
         # モデルオブジェクトを取得
         user = get_object_or_404(MyUser, pk=pk)
+        if user != request.user:
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # モデルオブジェクトを作成
         serializer = UserSerializer(
             instance=user, data=request.data, partial=True)
@@ -92,6 +98,8 @@ class UserRetrieveUpdateDestroyAPIView(views.APIView):
 
         # モデルオブジェクトを取得
         user = get_object_or_404(MyUser, pk=pk)
+        if user != request.user:
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # モデルオブジェクトを削除
         user.delete()
         # レスポンスオブジェクトを返す
@@ -102,6 +110,7 @@ class RoomAPIView(views.APIView):
     """
     ルームモデルの取得(一覧)・登録APIクラス
     """
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get(self, request, *args, **kwargs):
         """
@@ -109,7 +118,7 @@ class RoomAPIView(views.APIView):
         """
 
         # モデルオブジェクトの一覧を取得
-        room_list = Room.objects.all()
+        room_list = Room.objects.filter(is_private=False)
         # シリアライザオブジェクトを作成
         serializer = RoomSerializer(instance=room_list, many=True)
         # レスポンスオブジェクトを返す
@@ -134,6 +143,7 @@ class RoomRetrieveUpdateDestroyAPIView(views.APIView):
     """
     ルームモデルの取得(詳細)・更新・一部更新・削除APIクラス
     """
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get(self, request, pk, *args, **kwargs):
         """
@@ -154,6 +164,8 @@ class RoomRetrieveUpdateDestroyAPIView(views.APIView):
 
         # モデルオブジェクトを取得
         room = get_object_or_404(Room, pk=pk)
+        if not room.hosts.filter(pk=request.user.id).exists():
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # シリアライザオブジェクトを作成
         serializer = RoomSerializer(instance=room, data=request.data)
         # バリデーション
@@ -170,6 +182,8 @@ class RoomRetrieveUpdateDestroyAPIView(views.APIView):
 
         # モデルオブジェクトを取得
         room = get_object_or_404(Room, pk=pk)
+        if not room.hosts.filter(pk=request.user.id).exists():
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # モデルオブジェクトを作成
         serializer = RoomSerializer(
             instance=room, data=request.data, partial=True)
@@ -182,11 +196,13 @@ class RoomRetrieveUpdateDestroyAPIView(views.APIView):
 
     def delete(self, request, pk, *args, **kwargs):
         """
-        ロームモデルの削除APIに対応するハンドラメソッド
+        ルームモデルの削除APIに対応するハンドラメソッド
         """
 
         # モデルオブジェクトを取得
-        room = get_object_or_404(MyUser, pk=pk)
+        room = get_object_or_404(Room, pk=pk)
+        if not room.hosts.filter(pk=request.user.id).exists():
+            raise exceptions.AuthenticationFailed('Unauthorized access')
         # モデルオブジェクトを削除
         room.delete()
         # レスポンスオブジェクトを返す
